Inline methods that fold to a constant before analysis
The plugin NativeImageInlineDuringParsingPlugin analyses the graph for the resolved Java method and specifies
what should to be inlined during graph parsing before the static analysis. Plugin is searching for the methods
that folds to a constant.

native-image option for this plugin is called InlineBeforeAnalysis

remove duplicates of the method getCallingContextSubset